pyproject.toml

[tool.poetry]
name = "my_scraping_tool"
version = "0.1.0"
description = ""
authors = ["koba-masa <kobayashi_mshr@koba-masa.com>"]

[tool.poetry.dependencies]
python = "^3.9.4"
requests = "^2.25.1"
bs4 = "^0.0.1"
pytest = "^6.2.4"
python-dotenv = "^0.18.0"
pyyaml = "^6.0"
selenium = "^4.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
